backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
supabase==2.3.7
transformers==4.37.2
torch==2.2.0
httpx==0.26.0
slowapi==0.1.9
python-jose[cryptography]==3.3.0
pytest==8.0.0
numpy==1.26.4



